pyproject.toml
[tool.poetry]
name = "wagtail_non_admin_draftail"
version = "0.18.0"
description = "You can now use Wagtail Draftail editor on non-admin pages"
authors = ["Tim Kamanin <tim@timonweb.com>"]
license = "MIT"
homepage = "https://timonweb.com"
repository = "https://github.com/timonweb/wagtail-non-admin-draftail"

[tool.poetry.dependencies]
python = "^3.8.1"
wagtail = ">=5.0"

[tool.poetry.dev-dependencies]
pytest-django = "^4.4.0"
pytest-playwright = "^0.3.3"
django-managepy-anywhere = "^2.0"
pytest-faker = "^2.0.0"
pytest-env = "^0.8.2"
pre-commit = "^3.3.3"
flake8 = "^6.0.0"
black = "^23.3.0"
isort = "^5.12.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

[tool.black]
line-length = 88
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 88
known_third_party = "django,wagtail,pytz,redis,requests,rest_framework"
